clk_mali is useless it seems.
So according to mmind00 on the rockchip-linux IRC channel, defining a
new clock name for the GPU node is useless, any dependency to such
symbols should be removed.

So it's removed.

The patch adding the clock name has been removed and the drivers
dependencies to this clock name have been patched.

See here for the quick discussion on this subject :
https://irclog.whitequark.org/linux-rockchip/2017-07-18

Signed-off-by: Myy <myy@miouyouyou.fr>