One last try for the Asus Tinkerboard patch
If *THAT* does not work for you, ask for a refund.

This is an almost exact replica of the ARMbian patch, which is known
to work.

The only differences are in the emergency reboot routine, where I added
one printk, to differentiate this part from the simple reboot routine,
and the fact that I avoid to dereference a potential NULL variable
if I'm not sure that we talking about the same board.

I did a 'meld' on the miniarm.dts and the tinker.dts :
THEY DEFINE THE SAME NODES THE SAME WAY !

The only differences are a few things are the identation and the props
placement.

Signed-off-by: Myy <myy@miouyouyou.fr>