The 4.15-rc3 with the ... r19p0 Mali Midgard GPU drivers
And without the Tinkerboard reboot patch, so Tinkerboard might fail to
reboot with this kernel. Be aware...

The GPU drivers were downgraded in order to deal with an issue that
is triggered by legacy drivers using the legacy interface to query the
drivers version. Since this legacy interface have been dropped in the
latest versions (notably the r21), I'll keep the r19p0 for a little
while...

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>