I'm still using that stupid reboot patch
The new one works for standard reboots. No questions asked.
However it does not for emergency reboots.

I have no idea why. Trying to figure out how reboot work in this case
is quite a feat.
Quite frankly, I had thought that the reboot handlers would still be
called in case of emergency but that does not seem to be the case...
There's a reboot handler provided by ARM but using it just throw me
back on UBoot, which starts to complain about not being able to set
the eMMC voltage.

So I really think that some part of the patches should be applied to
U-Boot directly, instead of trying to mess up with the kernel.

Still, if you don't care about reboot in case of kernel panics and
bad crashes, then use the new reboot patch. It's far cleaner and it
works for standard cases.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>