Bump to 4.18-rc5
This version includes a patch based on @mihailescu2m work, to deal
with kernel panic messages related to devfreq when using the Mali
Midgard drivers, on mainline kernels.

This new patch is due to uses of read_rcu_*lock , that are not
needed since the kernel 4.11-rc1.
https://github.com/torvalds/linux/commit/5b650b388844f26c61c70564865598836d05dcb3
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/drivers/base/power/opp/core.c?id=5b650b388844f26c61c70564865598836d05dcb3

@mihailescu2m original patch is here :
https://github.com/mihailescu2m/linux/commit/bbe73c3c1143e5991bdcaee3afaecf5c31af0647

And the issue fixed by this patch is documented here :
https://community.arm.com/graphics/f/discussions/10246/unofficial-patches-for-the-mali-midgard-kernel-drivers-r19p0-on-linux-4-17

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>