The 4.20-rc2... With sound DMA broken and semi-broken Mali drivers
Everything is fine...
Ugh....

It seems that the switch from vm_insert_pfn to vfm_insert_pfn was done
through multiple Linux versions and the Mali driver, being a OOT driver,
is the last one to use it.
On a related note, most drivers now use vm_fault_t as a type to report
MMAP fault issue, and vmf_insert_pfn uses the same return type too.
Now, most of the ".fault" function pointers implementations I found in
the kernel tree are very simple, compared to the huge function used by
the Mali drivers. So I have no idea about how to correctly reimplement
this...

With this patch, glmark-es2-drm and X11 work but with degraded
performances, since the memory allocation are still not done correctly.
I guess that I should take a look at the last Mali kernel drivers.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>