The v5.1-rc4
With... the old reboot patches.

Turns out they changed a few things in the RK3288.dtsi that made
me think : At last, they solved the Tinkerboard reboot issue in a
clean way !

But no...

I wonder if it's because they didn't send all the required patches
to the mainline kernel, something that tends to happen a bit
too often, or if it's due to not testing these patches, or maybe
these patches don't target the reboot issue at all...

Meanwhile, a veyron patch has been removed since the property removed
was not there anymore. I didn't check if it's due to fixing the bug
upstream, or due to unrelated changes though... Since I don't have
a Chromebook, I can't test how well/bad this influences the stability
on these platforms.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>