The 5.8 !
With the ability to set the GPU back to 500Mhz, after some Chromebook
dev remove the ability for every RK3288 user to do that, in order to
protect their broken baby.

I also removed the old, unused and broken patches.

I also fixed a little issue with RTL8723BS but... while the driver
works in station mode, the Access Point mode is broken as shit.
And the driver is really badly written...

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>